pyproject.toml
[project]
name = "moment-clone"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "inngest>=0.4.0",
    "google-cloud-speech>=2.26.0",
    "google-cloud-storage>=2.17.0",
    "google-cloud-texttospeech>=2.17.0",
    "google-genai>=1.0.0",
    "svix>=1.24.0",
    "stripe>=10.0.0",
    "python-multipart>=0.0.12",
    "PyJWT[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    "cachetools>=5.3.0",
    "email-validator>=2.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "types-cachetools>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "pipeline", "knowledge_base"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]  # B008: Depends/File/Header() in defaults is standard FastAPI pattern

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
